Refactor external Blueprint to it's own dir
The external blueprint can be refactored to it's own directory. This
will make it easier to pull back in future to a shared place, possible
utils.
