+++ app/external/views/external.py
@@ -1,4 +1,6 @@
-from .. import external
+from flask import Blueprint
+
+external = Blueprint('external', __name__)
 
 
 @external.route('/<framework_framework>/opportunities/<brief_id>')
