+++ app/main/__init__.py
@@ -5,8 +5,6 @@ from dmcontent.content_loader import ContentLoader
 main = Blueprint('buyers', __name__)
 dos = Blueprint('dos', __name__)
 
-external = Blueprint('external', __name__)
-
 content_loader = ContentLoader('app/content')
 content_loader.load_manifest('digital-outcomes-and-specialists', 'briefs', 'edit_brief')
 content_loader.load_manifest('digital-outcomes-and-specialists', 'brief-responses', 'output_brief_response')
@@ -34,5 +32,4 @@ def add_cache_control(response):
 
 from ..main import errors
 from .views import buyers as buyers_views
-from .views import external as external_views
 from .views import digital_outcomes_and_specialists
